refactor : apis 절대경로에 맞게 폴더명 수정
